Fix four crash/correctness bugs found by ASan/UBSan fuzzing and clang-tidy
- lib-image-load.c: AssignIMG() recursed once per mipmap level with no
  depth cap, driven by a raw 32-bit file field (BRRES TEX0 / TPL
  n_image); a corrupted texture with a huge count stack-overflowed the
  process. Clamp to the existing MAX_MIPMAPS bound.

- lib-gf3ds.c: IsGF1Motion()'s 'skel + 2 > size' bounds check ran in
  32-bit arithmetic, so an attacker-controlled skel offset near
  UINT32_MAX wrapped past the check and reached data[skel] far out of
  bounds (confirmed bus error via fuzzing). Do the addition in size_t.

- lib-image-load.c: the BCFNT/BFFNT font loader had the same 32-bit
  wraparound bug in two more checks (finf_len, ptr_glyph) -- found by
  inspection after the gf3ds case, before fuzzing reached this path.
  ptr_glyph's wrap would have deref'd a wild pointer.

- lib-szs-create.c: the string-pool scanner's zero-skip loop checked
  only 'off < subszs.size' before an unconditional 4-byte read, so a
  buffer whose size isn't a multiple of 4 let it read past the end
  (confirmed heap-buffer-overflow via fuzzing). Require a full 4 bytes
  to remain.

All four were confirmed with a live crash (three via ASan/UBSan
mutation-fuzzing of tests/fixtures + tests/mk_*.py output through
'wszst EXTRACT', one via clang-tidy's clang-analyzer, cross-checked by
hand) and verified fixed by re-running the crashing input. A fourth,
30804-run fuzz pass afterward found nothing further.

tests/regress.sh gives byte-identical PASS/FAIL/SKIP counts and the
identical set of failing test names before and after, so this
introduces no functional regressions.

Also adds .clang-tidy (clang-analyzer-*, bugprone-*, with two
extremely noisy/low-signal checks disabled) as project infrastructure
for future static-analysis passes.

+++ project/src/lib-gf3ds.c
@@ -402,7 +402,11 @@ int IsGF1Motion (const u8 *data, size_t size, size_t file_size)
 	u32 skel = rd_le32 (data + 4);
 	if (!skel)
 		return 0;
-	if (skel + 2 > size)
+	// 'skel' is an attacker-controlled u32 file offset; fuzzing found that a
+	// value near UINT32_MAX makes "skel + 2" wrap in 32-bit arithmetic to a
+	// small number that passes this check, then data[skel] reads far out of
+	// bounds. Do the addition in size_t so it can't wrap.
+	if ((size_t)skel + 2 > size)
 		return gf_truncated (size, file_size);
 	// skeleton starts with u8 bone count + u8 first-bone index
 	if (data[skel] < 1 || data[skel] > 200)
